requirements.txt
django==4.2.1
psycopg==3.1.8
djangorestframework=3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.10.0
drf-yasg==1.21.5
dj-rest-auth==4.0.0
Faker==18.7.0